chore: Reorganize data feed utils
